requirements/tests.txt
-r base.txt

coverage==7.5.3
pytest==8.2.1
model-bakery==1.18.1
pytest-django==4.8.0
pytest-cov==5.0.0
pycodestyle==2.11.1
tox==4.15.0

ipython==8.25.0
flake8==7.0.0
isort
black
